Micram Coding
There seems to be an issue with comms to the Hittite, it works over GPIB but only very slowly. I don't know if the issue is MATLAB or the Laptop so I'll change it up and write a simple script that only does the Twin Pulse output with the clock rate entered manually.